perf: qcom: Add L3 cache PMU driver.
This adds a new dynamic PMU to the Perf Events framework to program
and control the L3 cache PMUs in some Qualcomm Technologies SOCs.
The PMU exports formatting and event information to sysfs so it can
be used by the perf user space tools.

This commit is a squash of the following commits:
	I167d002a1f0f4210f5eb9be6eb47b43de238a162
	I54afe327ca4d09c3fffdac3a28b9fd60fc8cb4cf
	I5049bd08a474592e361eb242d7ba213d88e7361b
	I59887857af2cf30ee7f3b4cc2758559ed0729314
	I71216f4ac34df609eb54e804aff56f7ba6d8c49b
	Ifb0b5a8d93c81fa28e8dae5bccbe152effd46226
	I6118c8a4e7037355d537c27aa47dadf001329740
	I4ad26a7b7c9e916873af5dbbdf6b07a125c38ece
	Idb11211f44b996fd6213af4018763b4bacc07bcc
	I8b0c8ed4d21977f17c6f83474ce0c29eddd00f0b

Change-Id: I76b268ff6b6cb7ea0c47ce2700dc6f0758bc3dbb
Signed-off-by: Agustin Vega-Frias <agustinv@codeaurora.org>