KVM: arm64: introduce ITS emulation file with MMIO framework
The ARM GICv3 ITS emulation code goes into a separate file, but needs
to be connected to the GICv3 emulation, of which it is an option.
The ITS MMIO handlers require the respective ITS pointer to be passed in,
so we amend the existing VGIC MMIO framework to let it cope with that.
Also we introduce the basic ITS data structure and initialize it, but
don't return any success yet, as we are not yet ready for the show.

Change-Id: I553513152f63b4841f3a22670cde0298c2f9d4c4
Signed-off-by: Andre Przywara <andre.przywara@arm.com>
Git-commit: d463d9a295864b0afa82dfd0b53248ae214777a1
Git-repo: git://linux-arm.org/linux-ap.git
Signed-off-by: Shanker Donthineni <shankerd@codeaurora.org>