KVM: arm64: allow updates of LPI configuration table
The (system-wide) LPI configuration table is held in a table in
(guest) memory. To achieve reasonable performance, we cache this data
in our struct vgic_irq. If the guest updates the configuration data
(which consists of the enable bit and the priority value), it issues
an INV or INVALL command to allow us to update our information.
Provide functions that update that information for one LPI or all LPIs
mapped to a specific collection.

Change-Id: I7332080845596734918c36917a6acaaa1d9b9430
Signed-off-by: Andre Przywara <andre.przywara@arm.com>
Git-commit: ce7d2a08fe41c90fb7c51d2b101a00dd1dcc5cb9
Git-repo: git://linux-arm.org/linux-ap.git
Signed-off-by: Shanker Donthineni <shankerd@codeaurora.org>