KVM: arm64: handle ITS related GICv3 redistributor registers
In the GICv3 redistributor there are the PENDBASER and PROPBASER
registers which we did not emulate so far, as they only make sense
when having an ITS. In preparation for that emulate those MMIO
accesses by storing the 64-bit data written into it into a variable
which we later read in the ITS emulation.
We also sanitise the registers, making sure RES0 regions are respected
and checking for valid memory attributes.

Change-Id: I80e7f0f4858f497537bd72766308998e05af6df5
Signed-off-by: Andre Przywara <andre.przywara@arm.com>
Git-commit: 3439b2aeb028e7e6f7322b496ad6e3a7a56c26af
Git-repo: git://linux-arm.org/linux-ap.git
Signed-off-by: Shanker Donthineni <shankerd@codeaurora.org>