soc: qcom: Add IRQ combiner driver
Driver for interrupt combiners in the Top-level Control and Status
Registers (TCSR) hardware block in Qualcomm Technologies chips.

An interrupt combiner in this block combines a set of interrupts by
OR'ing the individual interrupt signals into a summary interrupt
signal routed to a parent interrupt controller, and provides read-
only, 32-bit registers to query the status of individual interrupts.
The status bit for IRQ n is bit (n % 32) within register (n / 32)
of the given combiner. Thus, each combiner can be described as a set
of register offsets and the number of IRQs managed.

The combiner driver is implemented as a gpio_chip to allow consumers
to find their IRQs using devm_gpiod_get_index/gpiod_to_irq under ACPI.
This is done because the ACPI core does not use the Resource Source
field on the Extended Interrupt Descriptor, which in theory could be
used to specify an alternative IRQ controller similar to the way
interrupt-parent is used in DT/OF. Under ACPI boot, consumers will
specify their IRQs as GpioInt resources, which can be found with
devm_gpiod_get_index, and then mapped to irqs by calling gpiod_to_irq.

This commit is a squash of the following change-ids:
	I92c5d3a62da96ceb307fa298648fef8f95ef9f0d
	I0d1a924e20190782f54dead5de1fc5a42d39b75a
	I34f269b061da92aa7ea5bbbb1dc9694cb8f3ac5d

Change-Id: I71ef752de9402f63fc77b88fe37958a2104cabb8
Signed-off-by: Agustin Vega-Frias <agustinv@codeaurora.org>