acpi: apei: panic OS with fatal error status block
Even if an error status block's severity is fatal, the kernel does not
honor the severity level and panic.

With the firmware first model, the platform could inform the OS about a
fatal hardware error through the non-NMI GHES notification type. The OS
should panic when a hardware error record is received with this
severity.

Call panic() after CPER data in error status block is printed if
severity is fatal, before each error section is handled.

Change-Id: I09c1b5348aeb36752b5b555f0d9db8c179f621b8
Signed-off-by: Jonathan (Zhixiong) Zhang <zjzhang@codeaurora.org>