irqchip: its-platform: DO NOT UPSTREAM: add ACPI support
This is temporary solution to enable MSI interrupts for the platform
devices such as HIDMA and SMMUv3. The requirement is to have an
msi-support and device-id fields in ACPI.

This code can be extended in the future to figure out this information
from the IORT table. We are not there yet.

(cherry picked from commit 95048ca8057d9205279af357dd9f729a6a5abb93)
(okaya:codeaurora.org: probing changes for 4.7)
Signed-off-by: Sinan Kaya <okaya@codeaurora.org>

Change-Id: Ic597174f9d6f08b6f1daf4d57961e91157f9743c