soc: qcom: Add kernel panic notifier
When a kernel panic occurs, trigger an interrupt to notify other subsystems
of the crash. A proper separation between the platform and the OS is
achieved by using an ACPI function notification. The notification is
propagated from the source (kernel) to the receiver (firmware or other
subsystem) through an ACPI function.

The platform device is only probed if a notification receiver has
registered the function in ACPI.

This commit is a squash of the following change-ids:
	I499084120cbc0eaebba610f76e7618e9fbec180f
	Iabbf212a0ee2bc19ef38b01451e95a4f6af25b94

Change-Id: Ib046cd2ac3c8873fbcf91347de73ba2435472093
Signed-off-by: Tyler Baicar <tbaicar@codeaurora.org>
Signed-off-by: Richard Ruigrok <rruigrok@codeaurora.org>
Signed-off-by: Austin Christ <austinwc@codeaurora.org>