perf: increase arm_pmu used_mask size
used_mask is used to detect column exclusion in
ARMv7 krait processors. To extend this to ARMv8,
where there are more possible events, increase the
size of used_mask.

The maximum number of events in a current ARMv8
processor is 24 (3 PMRESR registers with 8 events
in each), plus 9 architected counters, for a total
of 33.

Change-Id: Ic3546be916a80457acbd1bb31a136a973745aae7
Signed-off-by: Neil Leeder <nleeder@codeaurora.org>