dmaengine: qcom_hidma: release the descriptor before the callback
There is a race condition between data transfer is finished and self
test reclaims the resources and returns back to the operating system.
Because of this race condition, we are leaking descriptors.

Instead of calling the callback first and then releasing the memory,
this code is changing the order to return the descriptor back to the
free pool and then call the user provided callback.

Change-Id: I9ce5c1f2acf985c0811e9545f611b2ce52a3f277
Signed-off-by: Sinan Kaya <okaya@codeaurora.org>
Signed-off-by: Richard Ruigrok <rruigrok@codeaurora.org>