ACPI: Support for platform initiated graceful shutdown
This patch adds support for platform initited graceful shutdown as
described in sections 5.6.6(Table-143) and 6.3.5.1 of ACPI 6.1 spec

The OSPM will get a graceful shutdown request via a Notify operator
on \_SB device with a value of 0x81 per section 5.6.6. Following the
shutdown request from platform the OSPM needs to follow the
processing sequence as described in section 6.2.5.1.

Change-Id: I190c44fe83f25b8df88a14727bfbf5d1d10565b8
Signed-off-by: Prashanth Prakash <pprakash@codeaurora.org>