arm64: qcom_defconfig: enable the CPU PMU driver
The CPU PMU driver enables the use of the implementation-
specific events in Qualcomm Technologies processors. It
integrates these events into the perf events subsystem.

Enabling these events is needed to allow collection of performance
statistics which are unique to Qualcomm Technologies processors
and can be used to perform performance analysis either system-wide
or on individual applications.

Change-Id: Ieff8711c2d7105bce6107a9d8d36453095ac1562
Signed-off-by: Neil Leeder <nleeder@codeaurora.org>