drivers: acpi: iort: add node match function
Device drivers (eg ARM SMMU) need to know if a specific component
is part of the IORT table, so that kernel data structures are not
initialized at initcalls time if the respective component is not
part of the IORT table.

To this end, this patch adds a trivial function that allows detecting
if a given IORT node type is present or not in the ACPI table, providing
an ACPI IORT equivalent for of_find_matching_node().

Change-Id: Ibc86512e101def809653250e492efc2d5d577593
Signed-off-by: Lorenzo Pieralisi <lorenzo.pieralisi@arm.com>
Cc: Hanjun Guo <hanjun.guo@linaro.org>
Cc: Tomasz Nowicki <tn@semihalf.com>
Cc: "Rafael J. Wysocki" <rjw@rjwysocki.net>
Git-commit: 92a3d58be1a18f3ddf3ec341257902d80aeeeeaa
Git-repo: git://git.kernel.org/pub/scm/linux/kernel/git/lpieralisi/linux.git
[nwatters@codeaurora.org: resolve trivial merge conflicts]
Signed-off-by: Nate Watterson <nwatters@codeaurora.org>