trace, ras: add ARM processor error trace event
Currently there are trace events for the various RAS
errors with the exception of ARM processor type errors.
Add a new trace event for such errors so that the user
will know when they occur. These trace events are
consistent with the ARM processor error section type
defined in UEFI 2.6 spec section N.2.4.4.

Change-Id: Iaf3c1f099fb149ddff159ada07b73547df55421d
Signed-off-by: Tyler Baicar <tbaicar@codeaurora.org>