KVM: arm/arm64: VGIC: add refcounting for IRQs
In the moment our struct vgic_irq's are statically allocated at guest
creation time. So getting a pointer to an IRQ structure is trivial and
safe. LPIs are more dynamic, they can be mapped and unmapped at any time
during the guest's _runtime_.
In preparation for supporting LPIs we introduce reference counting for
those structures using the kernel's kref infrastructure.
Since private IRQs and SPIs are statically allocated, the reqcount never
drops to 0 at the moment, but we increase it when an IRQ gets onto a VCPU
list and decrease it when it gets removed.
This introduces vgic_put_irq(), which wraps kref_put and hides the
release function from the callers.

Change-Id: Ibd5cc3f2660eff51943af65ea66e4d017345bfe2
Signed-off-by: Andre Przywara <andre.przywara@arm.com>
Git-commit: a277cf6997ee587fef21b34d9b036b20593abdcf
Git-repo: git://linux-arm.org/linux-ap.git
Signed-off-by: Shanker Donthineni <shankerd@codeaurora.org>