PCI: Add parent device field to ECAM struct pci_config_window
Add a parent device field to struct pci_config_window.  The parent is not
saved now, but will be useful to save it in some cases.  For ACPI on ARM64,
it can be used to setup ACPI companion and domain.

Since the parent dev is in struct pci_config_window now, we need not pass
it to the init function as a separate argument.

Change-Id: I4e4ef0cb9e0b767fbb73dd860cf1db7b4d174986
Signed-off-by: Jayachandran C <jchandra@broadcom.com>
Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>
Acked-by: Lorenzo Pieralisi <lorenzo.pieralisi@arm.com>
Git-commit: 5c3d14f76fccd14f0882fe2a11e19534a11a1674
Git-repo: https://github.com/semihalf-nowicki-tomasz/linux
Signed-off-by: Shanker Donthineni <shankerd@codeaurora.org>