iommu/arm-smmu-v3: DO-NOT-UPSTREAM: introduce macros to improve log readability
This patch introduces a set of logging macros that are functionally
equivalent to dev_{err, etc..} except they identify the smmu using
a "log name" instead of the often cryptic name produced by dev_name().
This makes debugging issues on systems with many smmus far easier as
one need not waste time repeatedly matching device names with smmu
hardware instances.

Currently, an smmu's log_name is specified by setting a string entry
called "smmu-name" in the devices firmware properties. In the case
where no log_name is found for an smmu, dev_name() is used as a
fallback.

For an iort defined smmu, if no "smmu-name" property is found,
bits [47:16] of the smmu base address are used as the log_name.

Change-Id: I23175932a5b596cb80bb9c9d3ab9ce83925bdfdf
Signed-off-by: Nate Watterson <nwatters@codeaurora.org>