iommu/arm-smmu-v3: DO-NOT-UPSTREAM: introduce trace_groups debug option
When enabled, trace_groups will register an iommu notifier with
all of the default groups for smmu connected masters. This notifier
logs all events (eg: ADD/REMOVE, BIND/UNBIND) as they occur which
makes it a valuable tool for debugging issues at the interface
between the dma and iommu layers.

Change-Id: Iab5eba853c3cc216fd1f7ce041f27f3a4baf071e
Signed-off-by: Nate Watterson <nwatters@codeaurora.org>