Documentation/watchdog: use stdout instead of stderr in watchdog-test
The watchdog-test utility outputs all messages to stderr, even those
that are not error messages.  Output to stdout instead.

Instead of flushing the output after every write, just disabled
the output buffer.

Also display a dot for every ping of the watchdog, so that the user
knows that it's working.

Change-Id: I2ca0a9d98b996129a38ba31b93da18a1e4f21b84
Signed-off-by: Timur Tabi <timur@codeaurora.org>