arm64: Add back cpu reset routines
Commit 68234df4ea79 ("arm64: kill flush_cache_all()") removed the global
arm64 routines cpu_reset() and cpu_soft_restart() needed by the arm64
kexec and kdump support.  Add back a simplified version of
cpu_soft_restart() with some changes needed for kexec in the new files
cpu_reset.S, and cpu_reset.h.

When a CPU is reset it needs to be put into the exception level it had when
it entered the kernel. Update cpu_soft_restart() to accept an argument
which signals if the reset address should be entered at EL1 or EL2, and
add a new hypercall HVC_SOFT_RESTART which is used for the EL2 switch.

Change-Id: Ic8dfae668140121ed6804805249fe03357a7a44e
Signed-off-by: Geoff Levand <geoff@infradead.org>
Git-commit: 9e9cc6f355e353bba1ddcc1f98f9fba41f9dcbf2
Git-repo: git://git.kernel.org/pub/scm/linux/kernel/git/geoff/linux-kexec.git
Reviewed-by: James Morse <james.morse@arm.com>
Signed-off-by: Sameer Goel <sgoel@codeaurora.org>