iommu/iova: validate iova_domain input to put_iova_domain
Passing a NULL or uninitialized iova_domain into put_iova_domain
will currently crash the kernel when the unconfigured iova_domain
data members are accessed. To prevent this from occurring, this patch
adds a check to make sure that the domain is non-NULL and that the
domain granule is non-zero. The granule can be used to check if the
domain was properly initialized because calling init_iova_domain
with a granule of zero would have already triggered a BUG statement
crashing the kernel.

Change-Id: If3ef9960697ee0d9b9bb3f7bd15cdd955ea0567a
Signed-off-by: Nate Watterson <nwatters@codeaurora.org>