irqchip: refactor and add GICv3 definitions
arm-gic-v3.h contains bit and register definitions for the GICv3 and ITS,
at least for the bits the we currently care about.
The ITS emulation needs more definitions, so add them and refactor
the memory attribute #defines to be more universally usable.
To avoid changing all users, we still provide some of the old definitons
defined with the help of the new macros.

Change-Id: Idd68b4393425a297f0aa553b3a4babeb8c10e641
Signed-off-by: Andre Przywara <andre.przywara@arm.com>
Git-commit: 1b43913652cb7e67394f51989c9a73c190d8cde0
Git-repo: git://linux-arm.org/linux-ap.git
[shankerd@codeaurora.org: Fix merge conflicts in header file]
Signed-off-by: Shanker Donthineni <shankerd@codeaurora.org>