arm64: pmu: Add support for probing with ACPI
In the case of ACPI, the PMU IRQ information is contained in the
MADT table. Also, since the PMU does not exist as a device in the
ACPI DSDT table, it is necessary to create a platform device so
that the appropriate driver probing is triggered.

Change-Id: I9e8c5383b3720d9ffd61c5b5403b426d0b74e7a6
Signed-off-by: Mark Salter <msalter@redhat.com>
Patch-mainline: https://patchwork.kernel.org/patch/8574251/
Signed-off-by: Neil Leeder <nleeder@codeaurora.org>