arm64: perf: make functions non-static
Allow common perf utility functions to be used by other perf drivers.

A driver which extends the basic CPU PMU functionality in order to support
a different PMU, such as that on chips from Qualcomm Technologies, will
need to use some of the core perf-events functions. Exporting their
definitions so that they can be shared by additional drivers means that
this functionality does not have to be duplicated in those drivers.

Change-Id: I6166531dc535d0df233eb28c3e88b2d63e8d8d3e
Signed-off-by: Neil Leeder <nleeder@codeaurora.org>