soc: qcom: provide mechanism for drivers to access L2 registers
L2 registers are accessed using a select register and data
register pair. To prevent multiple concurrent writes to the
select register by independent drivers, the write to the
select register and the associated access of the data register
are protected with a lock. All drivers accessing the L2
registers use the set and get functions provided by
l2-accessors to ensure correct reads and writes to L2 registers.

Change-Id: I96f056a7a3c21b76a1b247ac81da9bc25ef2a8ff
Signed-off-by: Neil Leeder <nleeder@codeaurora.org>