arm64: call cpu_pm* APIs only on context losing states
LPI(Low Power Idle) tables contains necessary information to indicate
if an idle state is a context losing one or not. This patch adds a
workaround to avoid saving context on entering non-context losing idle
states.

No need to upstream this patch as the proper fix will be made in GICv3
by Sudeep(from ARM).

Change-Id: Ibe24c391dda5da70493c650d5eba908363a78c6a
Signed-off-by: Prashanth Prakash <pprakash@codeaurora.org>