!FIX: vGIC: vgic_v2_probe: change init order to allow clean exit
if the kvm_register_vgic_device() call fails, we have to properly roll
back all the previous initialisation steps. That doesn't sound easily
doable with the HYP I/O mapping, so let's just move this mapping to be
the last step in the sequence.

Change-Id: I3c84a708121cffaebc16a19986584aeacc606ad2
Signed-off-by: Andre Przywara <andre.przywara@arm.com>
Git-commit: ba81e7abebdfe9d67fb061c2da3c87eb4e87765f
Git-repo: git://linux-arm.org/linux-ap.git
Signed-off-by: Shanker Donthineni <shankerd@codeaurora.org>