arm64: mm: prevent spurious calls to do_iommu_attach
Prior to this patch, whenever __iommu_attach_notifier was called with
a BIND_DRIVER action, all of the devices in iommu_dma_masters would be
sent to do_iommu_attach whether or not they were ready to be switched
over to the iommu_dma_ops. In order to reduce the log spew this was
causing, this patch only sends masters to do_iommu_attach if they
match the device input to __iommu_attach_notifier.

Change-Id: Iedcdf9a824c1df466dd6928c875d0e12c0e6b6b7
Signed-off-by: Nate Watterson <nwatters@codeaurora.org>