iommu/arm-smmu-v3: keep smmu hw and sw coherency settings consistent
Currently, the arm-smmu-v3 driver programs all smmus to make coherent
accesses to their memory resident data structures. This is problematic
when a particular smmu has a coherent path to memory but the firmware
properties for the corresponding device instance indicate that it
should be configured as non-coherent. In such cases, smmu commands
will be written by software in normal non-cacheable memory and may
not be seen by smmu hardware as it fetches stale data from the cache
that had been allocated by a prior read from the smmu. To prevent
such scenarios from occurring, this patch configures the coherency
settings for a particular smmu to match those specified in firmware.

Change-Id: I72570dd4e07a1c533eb9204002aa5c14efba2bda
Signed-off-by: Nate Watterson <nwatters@codeaurora.org>