iommu/arm-smmu-v3: DO-NOT-UPSTREAM: create a unique group per master
Currently, smmu attached pci devices are being lumped into a single
group per root complex. Because of this, all of the devices on a
root complex must be unbound from their normal driver in the host
kernel and rebound to vfio-pci for any of them to be directly assigned
to a guest machine. To allow for more flexibility, this patch assigns
every pci attached master to a unique group as was already being done
for platform devices.

Change-Id: If9ff629066b9734bedef00945e829d5cd4f738cc
Signed-off-by: Nate Watterson <nwatters@codeaurora.org>