iommu: arm-smmu-v3: add ATS support
Adding support for the ATS protocol. The PCIe endpoint can cache the
SMMU mappings by performing an ATS request in its own cached (ATC). The
translation agent needs to inform endpoint about mappings when they are
invalidated.

There are two categories of invalidations happening inside the SMMUv3
driver. Invalidation by page range and context invalidation.

Whenever invalidation by page range is issued, an ATC invalidate request is
sent to each PCIe endpoint in the same SMMU domain.

When context invalidation is issued, same operation is done. This time the
size is sent as 52 and address is sent as 0. According to the SMMUv3
specification, this means invalidate all entries in the ATC.

Change-Id: I704c186e747574162a695d833375d7f5c5966e57
Signed-off-by: Sinan Kaya <okaya@codeaurora.org>